[fix] rename components to zaloopa
